# perf: improve codec performance