Closes: #XXX

## What is the purpose of the change

> Add a description of the overall background and high-level changes that this PR introduces

![image](https://github.com/Argus-Labs/world-engine/assets/4402697/bcd166bc-f23a-4a16-8347-60ae757c46be)

Our current codec is unoptimized. It uses the stdlib json that have low hanging performance improvement. We also used an encoder unnecessarily when we could just use marshal/unmarshal, leading to an additional object allocation.

By doing this, we are able to significantly speed up the codec while reducing memory foot print. See benchmark results below 

**Before**
```
goos: darwin
goarch: arm64
pkg: pkg.world.dev/world-engine/cardinal/ecs/codec
BenchmarkDecode
BenchmarkDecode       	 1786486	       710.6 ns/op	    1056 B/op	      10 allocs/op
BenchmarkDecode-5     	 1998490	       597.9 ns/op	    1056 B/op	      10 allocs/op
BenchmarkDecode-10    	 2008329	       592.2 ns/op	    1056 B/op	      10 allocs/op
BenchmarkEncode
BenchmarkEncode       	 7187788	       164.3 ns/op	     136 B/op	       3 allocs/op
BenchmarkEncode-5     	 7717502	       155.8 ns/op	     136 B/op	       3 allocs/op
BenchmarkEncode-10    	 7776729	       155.2 ns/op	     136 B/op	       3 allocs/op
```

**After**
```
goos: darwin
goarch: arm64
pkg: pkg.world.dev/world-engine/cardinal/ecs/codec
BenchmarkDecode
BenchmarkDecode       	10113783	       105.2 ns/op	      56 B/op	       2 allocs/op
BenchmarkDecode-5     	12597741	        94.95 ns/op	      56 B/op	       2 allocs/op
BenchmarkDecode-10    	12556930	        96.06 ns/op	      56 B/op	       2 allocs/op
BenchmarkEncode
BenchmarkEncode       	13534896	        90.63 ns/op	      56 B/op	       2 allocs/op
BenchmarkEncode-5     	15366384	        76.47 ns/op	      56 B/op	       2 allocs/op
BenchmarkEncode-10    	15823453	        81.75 ns/op	      56 B/op	       2 allocs/op
PASS
```

## Brief Changelog

- Use https://github.com/goccy/go-json which is a drop-in replacement of stdlib json
- Switch to marshal/unmarshal instead of encoder/decoder
- Add benchmarks

## Testing and Verifying

- Added benchmark